+++ app/sub-common/files/load-fox-file.js
@@ -1,0 +1,18 @@
+function loadFileSafely(tgtFilePath)
+{
+	var loadRes = null;
+	
+	try
+	{
+		loadRes = require(tgtFilePath);
+	}
+	catch(e)
+	{
+		loadRes = null;
+	}
+	
+	return loadRes;
+}
+
+
+module.exports = loadFileSafely;