+++ test-parts/part-e-rio_settings/settings/set-main.js
@@ -4,14 +4,28 @@ const expect = require("chai").expect;
 const commonPaths = require("../../../app/paths/files/app-paths");
 const foxPath = require(commonPaths.foxRelative);
 const commonFunctionsFile = require(commonPaths.testCommonFull);
-const settingsFile = require(foxPath.rioSettingsFile);
+const loadFoxFile = require(commonPaths.loadFox);
+const settingsFile = loadFoxFile(foxPath.rioSettingsFile);
 
 
 function testRemoteIoSettings()
 {
 	describe("Main", function()
 	{
+		checkFile();
 		checkExports();
+	});
+}
+
+
+function checkFile()
+{
+	describe("Remote IO Settings File", function()
+	{
+		it("Loaded", function()
+		{
+			commonFunctionsFile.testPresent(settingsFile);
+		});
 	});
 }