+++ test-parts/part-e-rio_settings/settings/set-props.js
@@ -4,7 +4,8 @@ const expect = require("chai").expect;
 const commonPaths = require("../../../app/paths/files/app-paths");
 const foxPath = require(commonPaths.foxRelative);
 const commonFunctionsFile = require(commonPaths.testCommonFull);
-const settingsFile = require(foxPath.rioSettingsFile);
+const loadFoxFile = require(commonPaths.loadFox);
+const settingsFile = loadFoxFile(foxPath.rioSettingsFile);
 
 
 function testRemoteIoProperties()